requirements.txt
# Core Framework Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP and API Requests
requests==2.31.0
httpx==0.26.0

# Google ADK (Generative AI SDK)
google-generativeai==0.3.2

# MCP SDK (Model Context Protocol)
fastmcp>=0.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
typing-extensions==4.9.0

